implement entire migRefStats, except live migbands tracking and banch addition
