+++ src/McRefCommon.h
@@ -15,6 +15,8 @@ double calculateCoalStats(double *elapsed_times, int *num_lineages, int size);
 
 int isLeafPop(int pop);
 
+int isLeafNode(int node, int gen);
+
 int areChildrenLeaves(int pop);
 
 int isAncestralTo(int father, int son);
@@ -45,5 +47,8 @@ int getNodePop(int nodeId, int gen);
 
 LikelihoodNode *getNode(int nodeId, int gen);
 
+int **lca_pops; // "cache" of LCA of all pops. for example, lca_pops[1][2] is the lca of pops 1 & 2. Filled once by computeLca().
+void computeLcas();
+int migLcaPop(int nodeId, int gen, int defaultLcaPop);
 
 #endif
