+++ src/mcref/RefMigStats.h
@@ -1,13 +1,17 @@
 #ifndef G_PHOCS_REFMIGSTATS_H
 #define G_PHOCS_REFMIGSTATS_H
 
+void calculateReferenceMigrationStats();
+
+void initRefMigStats();
+
+int calculateReferenceGenMigStats(int nodeId, int gen);
+
+// ============================
+
 void allocateRefMigStatsMem();
 
 void printRefMigStatsHeader(FILE *file);
-
-void calculateReferenceMigrationStats();
-
-void initRefMigStats() ;
 
 void printReferenceMigrationStats(int iteration, FILE *file);
